[AIRFLOW-6371] Move dag tests from test_core to test_dag (#6927)
