+++ tests/jobs/test_local_task_job.py
@@ -17,8 +17,8 @@
 # specific language governing permissions and limitations
 # under the License.
 #
-
 import multiprocessing
+import os
 import time
 import unittest
 
@@ -34,11 +34,11 @@ from airflow.utils import timezone
 from airflow.utils.db import create_session
 from airflow.utils.net import get_hostname
 from airflow.utils.state import State
-from tests.test_core import TEST_DAG_FOLDER
 from tests.test_utils.db import clear_db_runs
 from tests.test_utils.mock_executor import MockExecutor
 
 DEFAULT_DATE = timezone.datetime(2016, 1, 1)
+TEST_DAG_FOLDER = os.environ['AIRFLOW__CORE__DAGS_FOLDER']
 
 
 class TestLocalTaskJob(unittest.TestCase):
